pyproject.toml
[project]
name = "binance-flow-history"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-binance>=1.0.19",
]
